Rename decorators module to deprecation
