+++ py-polars/polars/utils/deprecation.py
